Feature/apl 575 sharding recovery

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/commands/UpdateSecondaryIndexCommand.java
@@ -26,11 +26,11 @@ public class UpdateSecondaryIndexCommand implements DataMigrateOperation {
     private DataTransferManagementReceiver dataTransferManagement;
     private List<String> tableNameList;
     private int commitBatchSize;
-    private long snapshotBlockHeight;
+    private int snapshotBlockHeight;
     private Set<Long> dbIdsExclusionList;
 
     public UpdateSecondaryIndexCommand(DataTransferManagementReceiver dataTransferManagement,
-                             int commitBatchSize, long snapshotBlockHeight, List<String> tableNameList, Set<Long> dbIdsExlusionList) {
+                             int commitBatchSize, int snapshotBlockHeight, List<String> tableNameList, Set<Long> dbIdsExlusionList) {
         this.dataTransferManagement = Objects.requireNonNull(dataTransferManagement, "dataTransferManagement is NULL");
         this.snapshotBlockHeight = snapshotBlockHeight;
         this.commitBatchSize = commitBatchSize <= 0 ? DEFAULT_COMMIT_BATCH_SIZE : commitBatchSize;
@@ -39,7 +39,7 @@ public class UpdateSecondaryIndexCommand implements DataMigrateOperation {
     }
 
     public UpdateSecondaryIndexCommand(DataTransferManagementReceiver dataTransferManagement,
-                             long snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
+                             int snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
         this(dataTransferManagement,  DEFAULT_COMMIT_BATCH_SIZE, snapshotBlockHeight, dbIdsExclusionList);
         tableNameList.add(BLOCK_INDEX_TABLE_NAME);
         tableNameList.add(TRANSACTION_SHARD_INDEX_TABLE_NAME);
@@ -49,7 +49,7 @@ public class UpdateSecondaryIndexCommand implements DataMigrateOperation {
     public UpdateSecondaryIndexCommand(
             DataTransferManagementReceiver dataTransferManagement,
             int commitBatchSize,
-            Long snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
+            int snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
         this(dataTransferManagement, commitBatchSize, snapshotBlockHeight, null, dbIdsExclusionList);
     }
